refactor: split RunTimeSettings functions into runtime_settings.go
Move RunTimeSettings-related pure functions from oviewer.go to a
dedicated oviewer/runtime_settings.go, and move corresponding tests
from oviewer_test.go to oviewer/runtime_settings_test.go.

No behavior changes.

Moved functions:
- NewRunTimeSettings
- NewStyle
- setOldStyle
- setOldPrompt
- updateRunTimeSettings
- updatePromptConfig
- updateRuntimeStyle
